+++ src/main/java/sptech/elderly/web/dto/usuario/ClienteOutput.java
@@ -10,12 +10,12 @@ import java.time.LocalDate;
 /**
  * DTO for {@link sptech.elderly.entity.UsuarioEntity}
  */
-public record ClienteOutput(@NotNull Integer id,
+public record ClienteOutput(@NotNull Long id,
                             @NotBlank String nome,
                             @NotBlank String email,
                             @NotBlank String documento,
                             @NotNull LocalDate dataNascimento,
                             @NotBlank String biografia,
-                            Integer genero,
+                            Long genero,
                             EnderecoOutput endereco) implements Serializable {
 }
